feat: restructure dirs
