+++ dezero/core/__init__.py
@@ -1,0 +1,3 @@
+from dezero.core.core import Function, Variable
+
+__all__ = ["Variable", "Function"]
