--- dezero/functions/function.py
@@ -1,18 +1,0 @@
-import numpy as np
-
-from dezero.variable import Variable
-
-
-class Function:
-    def __call__(self, input: Variable) -> Variable:
-        x = input.data
-        y = self.forward(x)
-        output = Variable(y)
-        self.input = input
-        return output
-
-    def forward(self, x: np.ndarray) -> np.ndarray:
-        raise NotImplementedError()
-
-    def backward(self, gy: np.float64) -> np.float64:
-        raise NotImplementedError()
