+++ tests/functions/test_connect_functions.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
+from dezero.core import Variable
 from dezero.functions import Exp, Square
-from dezero.variable import Variable
 
 
 class TestConnectFunctions:
